Modifications qui seront validées : 	modifié :         bin/main 	modifié :         makefile 	supprimé :        src/Brick.cpp 	supprimé :        src/Brick.h 	modifié :         src/Car.cpp 	modifié :         src/Car.h 	modifié :         src/Graphic.cpp 	modifié :         src/layer.cpp 	modifié :         src/layer.h 	modifié :         src/main.cpp 	modifié :         src/nNetwork.cpp 	modifié :         src/nNetwork.h

+++ src/layer.cpp
@@ -17,7 +17,7 @@ Layer::Layer(int const& m_nbIn, int const& m_nbOut): nbIn(m_nbIn),nbOut(m_nbOut)
 {
     randomFilling();
 }
-Layer::Layer(vector<int> const& v): nbIn(1),nbOut(v.size())
+Layer::Layer(vector<float> const& v): nbIn(1),nbOut(v.size())
 {
     randomFilling();//Clean this
     for(int i=0; i<(int)v.size(); i++)
